eme: remove `IMediaKeySessionContext` type
